third_party/blink/web_tests/external/wpt/css/css-color-adjust/meta/meta-color-scheme-first-valid-applies-expected.txt

This is a testharness.js-based test.
FAIL Tree order decides which meta color-scheme applies. assert_equals: expected (string) "dark" but got (undefined) undefined
Harness: the test ran to completion.
